:memo: add instructions to enable daily game
To play the current day's puzzle, the application needs to determine
the player's timezone offset. The `time` crate considers this
computation "unsound". But the way this application uses it, it's
safe.

This commit adds instructions to make it build a version so you can
play daily -- without having to keep track of the day number.